FTHISSHIT

+++ router/sr_router.h
@@ -91,7 +91,7 @@ void sr_arphandler (struct sr_instance* sr, uint8_t * packet/* lent */, unsigned
 void sr_iphandler (struct sr_instance* sr, uint8_t * packet/* lent */, unsigned int len, char* interface/* lent */);
 
 void create_ethernet_header (sr_ethernet_hdr_t * eth_hdr, uint8_t * new_packet, uint8_t* src_eth_addr, uint8_t *dest_eth_addr);
-void create_arp_header (sr_arp_hdr_t* arp_hdr, uint8_t* new_packet, uint8_t src_ip, uint8_t dest_ip, unsigned char *sha, unsigned char *tha);
+void create_arp_header (sr_arp_hdr_t* arp_hdr, uint8_t* new_packet, uint32_t src_ip, uint32_t dest_ip, unsigned char *sha, unsigned char *tha);
 void create_ip_header (sr_ip_hdr_t *ip_hdr, uint8_t* new_packet);
 void create_icmp_header (uint8_t* new_packet, uint8_t type, unsigned int code, int len);
